+++ App/canary/events/messageCreate.js
@@ -1,0 +1,9 @@
+const { Events } = require('discord.js');
+
+module.exports = {
+    name: Events.MessageCreate,
+    async execute(message) {
+        if (message.author.bot) return;
+        if (message.content.toLowerCase() === 'hi') await message.channel.send('hi');
+    },
+};