+++ App/src/commands/context-menu/user.js
@@ -1,5 +1,4 @@
 const { ContextMenuCommandBuilder, ApplicationCommandType, EmbedBuilder } = require('discord.js');
-const fs = require('fs');
 
 module.exports = {
     category: 'context-menu',
@@ -8,54 +7,14 @@ module.exports = {
         .setType(ApplicationCommandType.User)
         .setDMPermission(false),
     async execute(interaction) {
-        const jsonPath = './../../database/data.json';
         const user = interaction.targetUser;
         const member = interaction.targetMember;
-
-        // Read JSON file
-        let users = [];
-        try {
-            const data = fs.readFileSync(jsonPath, 'utf8');
-            users = JSON.parse(data);
-        }
-        catch (err) {
-            console.error('Error reading data.json:', err);
-            return interaction.reply({
-                content: 'There was an error reading the database.',
-                ephemeral: true,
-            });
-        }
-
-        // Check if the user is already registered
-        const userId = user.id.toString();
-        const userExists = users.find(u => u.user === userId);
-
-        // If the user is not registered, send a message to register
-        if (!userExists) return interaction.reply({ content: 'This user is not registered yet.', ephemeral: true });
-
-        // Embed color based on user role
-        let embedColor = 'Default';
-        const pirate = '1205938469797625947';
-        if (member.roles.cache.has(pirate)) embedColor = 0x71368A;
-        const captain = '1259986911460851783';
-        if (member.roles.cache.has(captain)) embedColor = 0x9B59B6;
-        const serverBooster = '1207653983850594335';
-        if (member.roles.cache.has(serverBooster)) embedColor = 0xF47FFF;
-        const treasurerOfTheNight = '1205938707442434118';
-        if (member.roles.cache.has(treasurerOfTheNight)) embedColor = 0x010000;
-        const serverOwner = '1205588374354796574';
-        if (member.roles.cache.has(serverOwner)) embedColor = 0xF1C40F;
-
-        // Create the embed with the user information
         const embed = new EmbedBuilder()
-            .setColor(embedColor)
             .setThumbnail(`${user.displayAvatarURL()}`)
             .addFields(
-                { name: 'User:', value: `${user}`, inline: true },
-                { name: 'Karma:', value: `:performing_arts: ${userExists.karma}`, inline: true },
-                { name: 'Coins:', value: `:coin: ${userExists.coins}`, inline: true },
+                { name: 'User:', value: `${user.username}`, inline: true },
                 { name: 'ID:', value: `${user.id}`, inline: true },
-                { name: 'Joined at:', value: `${member.joinedAt}`, inline: true },
+                { name: 'Joined at:', value: `${member.joinedAt}`, inline: false },
             );
         return interaction.reply({ embeds: [embed], ephemeral: true });
     },