+++ App/src/commands/fun/ping.js
@@ -2,14 +2,12 @@ const { SlashCommandBuilder } = require('discord.js');
 
 module.exports = {
     category: 'fun',
+    // 60 seconds
+    cooldown: 60,
     data: new SlashCommandBuilder()
         .setName('ping')
-        .setDescription('Replies with pong!'),
+        .setDescription('Replies with Pong!'),
     async execute(interaction) {
-        // const sent = await interaction.reply({ content: 'Pinging...', fetchReply: true });
-        const sent = await interaction.reply({ content: 'pinging...', fetchReply: true });
-        // interaction.editReply(`Roundtrip latency: ${sent.createdTimestamp - interaction.createdTimestamp}ms`);
-        interaction.editReply(`:ping_pong: pong **${sent.createdTimestamp - interaction.createdTimestamp}**ms`);
-        // return interaction.reply('Pong!');
+        await interaction.reply({ content: 'Pong!', ephemeral: true });
     },
 };