+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: ba7b25d - v154
+| Author: FJrodafo - Sat, 10 Aug 2024 14:04:04 +0100
+| 
 * Commit: db72b38 - v153
 | Author: FJrodafo - Sat, 10 Aug 2024 13:41:16 +0100
 |